requirements.txt
DateTime==4.2
dj-database-url==0.5.0
Django==2.0.5
django-cors-middleware==1.3.1
django-extensions==2.0.7
django-filter==1.1.0
django-filters==0.2.1
django-heroku==0.3.1
django-utils==0.0.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.3
djangotoolbox==1.8.0
gunicorn==19.8.1
Pillow==5.1.0
psycopg2==2.7.4
Pygments==2.2.0
pyjade==4.0.0
PyJWT==1.6.1
pytz==2018.4
six==1.11.0
whitenoise==3.3.1
zope.interface==4.5.0
